Fix changes the merge with master missed
Because molecules/server/__init__.py was moved to a
different directory and was changed substantially, git
thought it was deleted, and did not merge it correctly
via the merge with master, but created a new file instead.

This fixes the issues. And it moves the jena.py file to
the new directory.
